book-store: use integer number of cents, not floating-point numbers

+++ exercises/book-store/src/lib.rs
@@ -1,3 +1,3 @@
-pub fn lowest_price(_: &[usize]) -> f64 {
+pub fn lowest_price(_: &[usize]) -> usize {
     unimplemented!()
 }
